Prevent overwriting of existing file
If the remediated name for a file matches the name of another file, the latter file
would be overwritten when the first file is renamed. A check was added to make sure
this does not happen. Instead, an error message will be logged, and the file name
will remain unchanged; the name will need to be remediated by hand.